split out into hpp and cpp files
