+++ .gitignore
@@ -102,3 +102,5 @@ typings/
 build
 prebuilds
 package-lock.json
+compile_commands.json
+.cache
