+++ src/metrics/Collector.cpp
@@ -1,0 +1,7 @@
+#include "Collector.hpp"
+
+namespace datadog {
+  uint64_t Collector::time_to_micro(uv_timeval_t timeval) {
+    return timeval.tv_sec * 1000 * 1000 + timeval.tv_usec;
+  }
+}
