+++ src/metrics/Collector.hpp
@@ -12,8 +12,4 @@ namespace datadog {
     protected:
       virtual uint64_t time_to_micro(uv_timeval_t timeval);
   };
-
-  uint64_t Collector::time_to_micro(uv_timeval_t timeval) {
-    return timeval.tv_sec * 1000 * 1000 + timeval.tv_usec;
-  }
 }
