+++ src/metrics/Process.hpp
@@ -15,18 +15,4 @@ namespace datadog {
       bool enabled_;
       uv_rusage_t usage_;
   };
-
-  void Process::inject(Object carrier) {
-    uv_rusage_t usage;
-    uv_getrusage(&usage);
-
-    Object cpu;
-
-    cpu.set("user", time_to_micro(usage.ru_utime) - time_to_micro(usage_.ru_utime));
-    cpu.set("system", time_to_micro(usage.ru_stime) - time_to_micro(usage_.ru_stime));
-
-    carrier.set("cpu", cpu);
-
-    usage_ = usage;
-  }
 }
